Adding configuration for all indicators (#446)
**Included in this PR:**

- All indicators should use an object configuration per the feature
request described in this issue: Closes
https://github.com/cinar/indicatorts/issues/445
- Introduced shorter names for most functions (i.e. `awesomeOscillator`
is now exported as `ao` and `awesomeOscillator`. All internal usage uses
the abbreviated name if one exists. Abbreviations we're borrowed from
other similar libraries.
- Code coverage for indicators was brought up significantly. Strategy
tests were deemed out of scope for this PR.
- Documentation was altered accordingly

**Considerations:**

- I did not place deprecation messages in this PR because, in all
honesty, it's probably better reviewed as a major version bump. The
amount of messages we'd have to introduce into the codebase would be
significant.
- One other consideration is that I added a number of default values,
primarily for periods. I don't know what a lot of default values should
be - so I'll be seeking guidance on what you want those values to be.
- I can't vouch for the accuracy of the numbers in the tests. However,
since this PR doesn't alter the integrity of the functions themselves,
the only way they could be wrong is if the original logic was wrong to
begin with. With that said, it wouldn't be a bad time to verify these
values against other reputable libraries.
- I noticed that sometimes test data was the same, even when passing a
different `period`. This may be due to low quality mock data, but I'm
not sure why this is.